# docs/research/main_specs/1.0.0-mvp/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Helix OTA Control Plane API
  version: 1.0.0-MVP
  summary: REST API for the Helix OTA modular-monolith control plane (Android 15 first).
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  description: >
    Control-plane endpoints (auth, devices, artifacts, releases, deployments, client
    update-check + telemetry). JSON responses negotiate Brotli/gzip; the referenced
    artifact-byte path is served identity-encoded, ZIP_STORED, with HTTP Range support.
    Response-header conventions (applied to every negotiated JSON response, not repeated per
    operation): 'Vary: Accept-Encoding' is always set (the Vary header component), and 'X-Request-Id'
    is always returned. Rate-limited routes additionally return RateLimit-Limit / RateLimit-Remaining
    / RateLimit-Reset (draft IETF RateLimit fields); on 429 a Retry-After header is returned.
  contact:
    name: Helix OTA control-plane team
  # license: intentionally omitted — project license is UNVERIFIED in this revision.
servers:
  - url: https://{host}/api/v1
    description: Helix OTA control plane (HTTP/3 primary, HTTP/2 fallback).
    variables:
      host:
        default: helix.example
...
security:
  - bearerAuth: []
paths:
...
  /releases/{releaseId}:
    get:
      tags: [releases]
      operationId: getRelease
      summary: Read a single release.
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
      responses:
        '200':
          description: Release.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Release' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /deployments:
    post:
      tags: [deployments]
      operationId: createDeployment
      summary: Assign a release to a target set. MVP supports strategy=all-targets only.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeploymentCreate' }
      responses:
        '201':
          description: Deployment created.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
            RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deployment' }
        '200':
          description: Idempotent replay of a prior deployment (same Idempotency-Key).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deployment' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
    get:
      tags: [deployments]
      operationId: listDeployments
      summary: List active deployments.
      description: >
        Returns the set of currently-active deployments. Lets operators and agents enumerate what
        is live (and read each deployment_id) without guessing ids. RBAC: viewer/operator/admin.
      responses:
        '200':
          description: Active deployments.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeploymentList' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /deployments/{deploymentId}:
    get:
      tags: [deployments]
      operationId: getDeployment
      summary: Read a deployment with aggregate progress.
      parameters:
        - $ref: '#/components/parameters/DeploymentId'
      responses:
        '200':
          description: Deployment status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeploymentStatus' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /deployments/{deploymentId}/rollbacks:
    get:
      tags: [rollback]
      operationId: listRollbacks
      summary: List the rollback/abort history for a deployment.
      description: RBAC viewer+ (handlers_recall.go:handleListRollbacks).
      parameters:
        - $ref: '#/components/parameters/DeploymentId'
      responses:
        '200':
          description: Rollback history.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RollbackList' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /projects:
    post:
      tags: [projects]
      operationId: createProject
      summary: Create a project within the token's account.
      description: >
        RBAC operator/admin. The project is scoped to the account carried in the
        JWT token claim (handlers_project.go:handleCreateProject).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreate' }
      responses:
        '201':
          description: Project created.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectView' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [projects]
      operationId: listProjects
      summary: List projects (account-scoped, cursor-paginated).
      description: >
        RBAC viewer+. Returns projects for the token's account
        (handlers_project.go:handleListProjects).
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Projects list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectList' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /fabric/nodes/{nodeId}:
    get:
      tags: [fabric]
      operationId: getFabricNode
      summary: Get a fabric node by ID (admin only).
      description: RBAC admin only (handlers_fabric.go:handleGetFabricNode).
      parameters:
        - { name: nodeId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Fabric node.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FabricNodeView' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /fabric/targets:
    post:
      tags: [fabric]
      operationId: registerFabricTarget
      summary: Register a fabric target (admin only).
      description: RBAC admin only (handlers_fabric.go:handleRegisterFabricTarget).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FabricTargetRegister' }
      responses:
        '201':
          description: Target registered.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FabricTargetView' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    get:
      tags: [fabric]
      operationId: listFabricTargets
      summary: List fabric targets (admin only).
      description: RBAC admin only (handlers_fabric.go:handleListFabricTargets).
      responses:
        '200':
          description: Fabric targets list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FabricTargetList' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # --- Unversioned operational probes ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        OAuth2/JWT access token (admin/operator/viewer) or device-scoped token (role=device).
        RBAC enforced per-route; device tokens may act only on their own deviceId.
    provisioningToken:
      type: http
      scheme: bearer
      description: >
        Alternative auth for POST /devices/register only: a provisioning token that authorizes a
        device to enrol itself when no operator bearer is available. The global bearerAuth (operator
        role) is the primary path; this scheme is the alternative (either satisfies the operation).
  headers:
    XRequestId:
      description: Correlation id; mirrors error.request_id.
      schema: { type: string }
...
    RetryAfter:
      description: Seconds to wait before retrying (on 429).
      schema: { type: integer, minimum: 0 }
    RateLimitLimit:
      description: Request quota for the current window (draft IETF RateLimit field).
      schema: { type: integer, minimum: 0 }
    RateLimitRemaining:
      description: Requests remaining in the current window (draft IETF RateLimit field).
      schema: { type: integer, minimum: 0 }
    RateLimitReset:
      description: Seconds until the current rate-limit window resets (draft IETF RateLimit field).
      schema: { type: integer, minimum: 0 }
  parameters:
...
    ReleaseId:
      name: releaseId
      in: path
      required: true
      schema: { type: string, format: uuid }
    DeploymentId:
      name: deploymentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 255 }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
  responses:
    ValidationFailed:
      description: Request failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthenticated:
      description: Missing, invalid, or expired credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but lacks the required role or resource ownership.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: State conflict (duplicate, non-monotonic version, overlapping deployment).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    RateLimited:
      description: Rate limit exceeded.
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    SemVer:
      type: string
      pattern: '^\d+\.\d+\.\d+$'
      examples: ['1.1.0']
    Os:
      type: string
      enum: [android]
      description: MVP targets Android 15 first; Linux/universal is a later phase.
...
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              enum:
                - UNAUTHENTICATED
                - FORBIDDEN
                - NOT_FOUND
                - VALIDATION_FAILED
                - CONFLICT
                - UNSUPPORTED_MEDIA_TYPE
                - PAYLOAD_TOO_LARGE
                - RATE_LIMITED
                - SIGNATURE_INVALID
                - HASH_MISMATCH
                - VERSION_NOT_MONOTONIC
                - INTERNAL
            message: { type: string }
            request_id: { type: string }
            details:
              type: array
              items:
                type: object
                additionalProperties: false
                properties:
                  field: { type: string }
                  issue: { type: string }
...
    Release:
      type: object
      required: [release_id, artifact_id, version, os, target_model, status, created_at]
      additionalProperties: false
      properties:
        release_id: { type: string, format: uuid }
        artifact_id: { type: string, format: uuid }
        version: { $ref: '#/components/schemas/SemVer' }
        os: { $ref: '#/components/schemas/Os' }
        target_model: { type: string }
        status: { type: string, enum: [published, archived] }
        notes: { type: string }
        created_at: { type: string, format: date-time }
...
    DeploymentCreate:
      type: object
      required: [release_id, strategy]
      additionalProperties: false
      properties:
        release_id: { type: string, format: uuid }
        strategy:
          type: string
          const: all-targets
          description: MVP supports only 'all-targets'; staged rollout is 1.0.1 (ADR-0003).
        group: { type: string, description: Optional narrowing to a device group. }
    Deployment:
      type: object
      required: [deployment_id, release_id, strategy, status, created_at]
      additionalProperties: false
      properties:
        deployment_id: { type: string, format: uuid }
        release_id: { type: string, format: uuid }
        strategy: { type: string, const: all-targets }
        group: { type: string }
        status: { type: string, enum: [active, completed, halted] }
        target_count: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
    DeploymentStatus:
      allOf:
        - { $ref: '#/components/schemas/Deployment' }
        - type: object
          additionalProperties: false
          properties:
            progress:
              type: object
              additionalProperties: false
              properties:
                pending: { type: integer, minimum: 0 }
                downloading: { type: integer, minimum: 0 }
                installed: { type: integer, minimum: 0 }
                succeeded: { type: integer, minimum: 0 }
                failed: { type: integer, minimum: 0 }
    DeploymentList:
      type: object
      required: [items]
      additionalProperties: false
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Deployment' }
        next_cursor:
          oneOf:
            - { type: string }
            - { type: 'null' }
...
    RollbackView:
      type: object
      required: [id, deployment_id, kind, created_at]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        deployment_id: { type: string, format: uuid }
        kind: { type: string, enum: [abort, rollback] }
        from_release_id: { type: string, format: uuid }
        to_release_id: { type: string, format: uuid }
        recall_deployment_id: { type: string, format: uuid, description: The new forward-fix deployment created by a recall. }
        reason: { type: string }
        triggered_by: { type: string, description: Token subject of the operator that triggered the recall. }
        details:
          type: object
          additionalProperties: { type: string }
        created_at: { type: string, format: date-time }
    RollbackList:
      type: object
      required: [items]
      additionalProperties: false
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/RollbackView' }

    # --- telemetry reads (handlers_telemetry.go) ---
...
    ProjectCreate:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: { type: string }
        description: { type: string }
        branch_id: { type: string, format: uuid, description: Optional parent branch for the project. }
...
    ProjectView:
      type: object
      required: [project_id, name, account_id, created_at]
      additionalProperties: false
      properties:
        project_id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        account_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ProjectList:
      type: object
      required: [items]
      additionalProperties: false
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ProjectView' }
        next_cursor:
          oneOf:
            - { type: string }
            - { type: 'null' }

    # --- project members (handlers_project_members.go) ---
...
    FabricNodeView:
      type: object
      required: [node_id, host, port, created_at]
      additionalProperties: false
      properties:
        node_id: { type: string, format: uuid }
        host: { type: string }
        port: { type: integer }
        role: { type: string }
        properties:
          type: object
          additionalProperties: { type: string }
        created_at: { type: string, format: date-time }
    FabricTargetRegister:
      type: object
      required: [target_model, os]
      additionalProperties: false
      properties:
        target_model: { type: string }
        os: { $ref: '#/components/schemas/Os' }
        os_version: { type: string }
        properties:
          type: object
          additionalProperties: { type: string }
    FabricTargetView:
      type: object
      required: [target_id, target_model, os, created_at]
      additionalProperties: false
      properties:
        target_id: { type: string, format: uuid }
        target_model: { type: string }
        os: { $ref: '#/components/schemas/Os' }
        os_version: { type: string }
        properties:
          type: object
          additionalProperties: { type: string }
        created_at: { type: string, format: date-time }
    FabricTargetList:
      type: object
      required: [items]
      additionalProperties: false
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/FabricTargetView' }

    # --- health probes ---
...